Enhance join_with_slash(), and move to misc.py
Fixes #121
